Add helpers file to clean up code
