+++ helpers.js
@@ -1,0 +1,86 @@
+const {
+  differenceInHours,
+  differenceInMinutes,
+  differenceInSeconds,
+  subHours,
+  subMinutes,
+  subSeconds
+} = require('date-fns');
+
+const getServerTime = function formatsLocalTimeToServerTimeUnformatted() {
+  //Current Date with Time
+  const localTime = new Date();
+  //Current time in milliseconds
+  const localTimeinMs = localTime.getTime();
+  //Difference between your timezone and UTC in hours
+  const localTimezoneOffset = localTime.getTimezoneOffset() / 60;
+  /**
+   * The above is just for context
+   * If you're at UTC +8 timezone, the offset is UTC +0 minus UTC +8
+   * Below is more on how to get what you need
+   * Pretty much you'll just need the server's timezoneoffset
+   * And use that to configure localTime
+   */
+  const serverTimezoneOffset = 4; //EDT offset
+  const timezoneDifference = localTimezoneOffset - serverTimezoneOffset;
+  const serverTime = localTimeinMs + timezoneDifference * 3600000; //serverTime in milliseconds
+
+  return serverTime;
+};
+
+/**
+ * date-fns doesn't let you format distanceInWordsStrict yet
+ * so can't return X hours, y mins, z seconds and can only get individually
+ * This function tries to achieve this
+ */
+const formatCountdown = function formatCountdownUsingDifference(serverTime, nextSpawnDate) {
+  let formattedCountdown = [];
+  let calculatedTime = nextSpawnDate;
+  /**
+   * First checks hours and adds to array if it's over zero
+   * If it is, substracts hours from calculatedTime
+   * Then checks minutes and does the same thing and so on
+   * Wrote additional logic since I'm such a grammar nazi
+   * */
+  const countdownHours = differenceInHours(calculatedTime, serverTime);
+  if (countdownHours > 0) {
+    calculatedTime = subHours(calculatedTime, countdownHours);
+
+    switch (countdownHours) {
+      case 1:
+        formattedCountdown.push(`${countdownHours} hour`);
+        break;
+      default:
+        formattedCountdown.push(`${countdownHours} hours`);
+    }
+  }
+
+  const countdownMinutes = differenceInMinutes(calculatedTime, serverTime);
+  if (countdownMinutes > 0) {
+    calculatedTime = subMinutes(calculatedTime, countdownMinutes);
+    switch (countdownMinutes) {
+      case 1:
+        formattedCountdown.push(`${countdownMinutes} minute`);
+        break;
+      default:
+        formattedCountdown.push(`${countdownMinutes} minutes`);
+    }
+  }
+
+  const countdownSeconds = differenceInSeconds(calculatedTime, serverTime);
+  if (countdownSeconds > 0) {
+    switch (countdownSeconds) {
+      case 1:
+        formattedCountdown.push(`${countdownSeconds} second`);
+        break;
+      default:
+        formattedCountdown.push(`${countdownSeconds} seconds`);
+    }
+  }
+  return `in ${formattedCountdown.join(' ')}`;
+};
+
+module.exports = {
+  getServerTime: getServerTime,
+  formatCountdown: formatCountdown
+};
